requirements.txt
flask
streamlit
google-generativeai
python-dotenv
requests
